// frontend/src/api/interface/setting.ts
export namespace Setting {
// ...
    export interface NodeItem {
        id: number;
        addr: string;
        status: string;
        version: string;
        isXpack: boolean;
        isBound: boolean;
        name: string;
        value?: string;
        message?: string;
    }
// ...
}